common/parser.go: Trim ASCII-NUL strings to exclude binary characters
This is for convenience. We're already ignoring whether it does or does
not have the NUL. Since having binary characters is likely an
undesireable flaw in the image and support for binary characters by EXIF
is probably just a vestigial, ridiculous feature, just stop short and
return the preceding characters if encountered. If we do actually need
the binary at some point in the future, we can circle back and
potentially add an option to do so.

Closes #55